deactivate single command
We want to be able to deactivate single commands. Yet, if we deactivate
a command of a project, then we have to deactivate all release commands
as well. Why? If we deactivate an update dependency command and we still
release then we might end up with two versions of a project on the
classpath of a dependent project. Thus:
- introduce marker interface ReleaseCommand
  - deactivate also such commands, regardless if they are deactivated
    or another command
  - deactivate dependent projects (since if we do not release, then they
    would wait forever)
- write corresponding spec