spec multi module with cyclic release dependency
One problem we did not address yet is the following. Submodule has a
dependency on a library, that library has a dependency on the multi
module project which the submodule is part of. This would actually be
fine if we do not release the multi-module project as a whole. Thus
we have a release cycle dependency which we have to warn the user about
and where we need to take special care:
- Submodule cannot keep the level in such a scenario because we cannot
  add the dependency to the multi-module project