provide possibility to disable release command
One might have projects which should not be released at all (because
they are never released or require a manual process etc.). Provide an
option where a user can specify a regex for which all matching projects'
release commands shall be deactivated. Moreover, it shall not be
possible that a user activates the command via GUI in such a case. Thus:
- introduce CommandState.Disabled
- adopt serialization
- adopt GUI accordingly, disable reactivation in JS, disable checkbox
  and characterise disabled checkboxes with CSS
- add option to main and gradle-task `html`
- adopt JenkinsReleasePlanCreator accordingly, log if a project's
  release command is deactivated