write regression for multi module has release-relation
In case a multi module does not yet have a relation to a dependency set
and a submodule has one, then we need to update the multi module in case
it is not on the same level as the submodule. We call this a release-
relation because we have to wait with the release of the multi-module
due to the dependency of the submodule.
Now the problem is, that we currently only check if a dependent has
already a relation to the dependency and skip updating commands in such
a case. We need to change that:
- check that we do already have at least one command which is waiting
  for the dependency, if not, then we have the above situation where
  we still need to add commands for the multi-module
- write regression spec