document limitation property defined elsewhere
Consider the following situation, submodule uses a property as version
which is defined in the multi-module. The multi-module itself does
not use the property at all. We currently do not track the property
back to the multi module and thus we don't know that this property
actually semantically defines a version of a dependency.