deactivate/disable only depending commands in ReleasePlanManipulator
Furthermore, we have a bug that only projects on the next level are
deactivated/disabled. This is due to a misconception that we can use the
ReleasePlan.iterator() from any given point and reach all dependents.
That is not the case due to fact that a dependent of the chosen point
might have a dependency which is not even in the dependent tree of the
chosen point which means we expect the dependent to be on a different
level. Short:
- remove iterator overload with a given entry point
- adjust ReleasePlanManipulator only deactivate depending commands and
  release commands (not all commands)
- disabling a command should only disable the command and the release
  command but deactivate depending commands.