allow to continue a succeeded process
A user might want to continue a process even when it succeeded
successfully. For instance, if one has deactivated a certain command,
triggered the release, waited for its completion and then modified
the corresponding project manually, then one might want to re-activate
the command and continue with the release-process. Thus:
- allow transition from Succeeded to Ready
- reactivate toggles after success
- reactivate the corresponding button
  - change title to "Continue: ..."
- remove notice about "Use a new pipeline..." from the success message