introduce a start-over button so that release can follow dry run
As a user I would like to perform a dry-run first and then a release
after a success but without the need to re-open the pipeline in order
that a) made changes to versions, settings etc. is not lost and b) I do
not have to search the the corresponding pipeline.
We already remember the initial json in the Config. Thus:
- add a start-over button so that one can reset the process
  - un-disable text fields and toggles
  - reset buttons (so that they can be used again)
  - reset commands to their initial state