take into account that queueing command could already be executed
We cannot reuse the QueuedItemBasedBuildNumberExtractor when we recover
an ongoing process, a command which is in state queueing respectively.
The reason is simple, it is very likely that the queued job is now
already executing (or has already finished). Thus we need to deal with
this, use the BuildHistoryBasedBuildNumberExtractor as fallback:
- delete extractBuildNumber.kt -> move logic back to JenkinsJobExecutor
- implement own extractBuildNumber in recover.kt