introduce CommandState StillQueueing
when we recover from an ongoing process and we find the item in the
queue but still no number, then it is wrong to just assume that we have
to extract it from the build history. The contrary, this way we might
analyse an old build and not the current one. Hence:
- introduce a new state so that we can re-poll a queued item
- adjust recover.kt accordingly, check if the a job is still queueing
- adjust the releaser accordingly
  - introduce JobExecutor::rePollQueueing