remove release command and update commands if synthetic root
Currently we show a release-command which is already succeeded. That's
kind of OK but might also confuse the user. It would be better if we
remove this release command from the synthetic root.
Moreover, we have to remove the update commands in the projects which
are dependents of the synthetic root (so the real root projects). We
cannot make an update for the synthetic root.